+++ R/tmb.R
@@ -230,6 +230,7 @@ h_mmrm_tmb_data <- function(
     nm = colnames(x_matrix)
   )
   qr_x_mat <- qr(x_matrix)
+  x_matrix_complete <- x_matrix
   if (qr_x_mat$rank < ncol(x_matrix)) {
     cols_to_drop <- utils::tail(qr_x_mat$pivot, ncol(x_matrix) - qr_x_mat$rank)
     if (identical(singular, "error")) {
@@ -246,8 +247,8 @@ h_mmrm_tmb_data <- function(
         " by using `accept_singular = TRUE`"
       )
     } else if (identical(singular, "drop")) {
-      assign_attr <- attr(x_matrix, "assign")
       contrasts_attr <- attr(x_matrix, "contrasts")
+      assign_attr <- attr(x_matrix, "assign")
       x_matrix <- x_matrix[, -cols_to_drop, drop = FALSE]
       x_cols_aliased[cols_to_drop] <- TRUE
       attr(x_matrix, "assign") <- assign_attr[-cols_to_drop]
@@ -298,6 +299,7 @@ h_mmrm_tmb_data <- function(
       data = data,
       x_matrix = x_matrix,
       x_cols_aliased = x_cols_aliased,
+      x_matrix_complete = x_matrix_complete,
       coordinates = coordinates_matrix,
       y_vector = y_vector,
       weights_vector = weights_vector,